Fix handling vbox machines with spaces in names
From Pedro Ribeiro, via tuxonice-devel:

Hi,

the current virtualbox script in hibernate-2.0 has a major flaw: if a
virtual machine has a space in its name, as for example "Debian
32-bit", the script will fail to suspend the virtual machine.

This is bad for two reasons:
- a virtual machine occupies a large amount of RAM, and hence it will
make hibernate/resume a lot longer
- in some cases, the vboxdrv kernel module will crash on resume,
possibly taking down the whole system with it

The attached patch extract the stop_vms() function from the Debian
/etc/init.d/virtualbox-ose script to correctly save and suspend all
the running VMs of all users.
On resume, all VMs will be restored to their running state.

I have tested the patch with several running VMs and it works correctly.

Regards,
Pedro

Signed-off-by: Pedro Ribeiro <pedrib@gmail.com>
Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>