Add more intelligent vbox module removal.
After suspending virtual box machines, we need to also unload the
kernel modules in order to have a successful hibernation cycle. Since
there may still be references to the modules for a short while after
the vms are stopped, we need to include a loop, periodically retrying
the unloading of the modules. Original work was by Pedro, slight
modifications by Nigel.

Signed-off-by: Pedro Ribeiro <pedrib@gmail.com>
Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>